creates header controller
